contador y crasheo
He intentado hacer el contador y e echo un bucle infinito sin querer

+++ PR_ZAXXON_GomezNatalia/Assets/scripts/Volumen.cs
@@ -30,7 +30,8 @@ public class Volumen : MonoBehaviour
     }
     public void SubirVolumen()
     {
-        volumen = (mySlider.value);
+        volumen = ((int)
+            mySlider.value);
         volumenText.text = "Volumen:" + volumen;
     }
 }
